Rename proxy package to cfauthproxy
