+++ cmd/cf-auth-proxy/main.go
@@ -14,9 +14,9 @@ import (
 
 	envstruct "code.cloudfoundry.org/go-envstruct"
 	"code.cloudfoundry.org/log-cache/internal/auth"
+	. "code.cloudfoundry.org/log-cache/internal/cfauthproxy"
 	"code.cloudfoundry.org/log-cache/internal/metrics"
 	"code.cloudfoundry.org/log-cache/internal/promql"
-	. "code.cloudfoundry.org/log-cache/internal/proxy"
 	logtls "code.cloudfoundry.org/log-cache/internal/tls"
 	"code.cloudfoundry.org/log-cache/pkg/client"
 )
