+++ internal/cfauthproxy/cf_auth_proxy.go
@@ -1,4 +1,4 @@
-package proxy
+package cfauthproxy
 
 import (
 	"log"
